misc/HtmlTestSuite/results/test-1001.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
 <head>
  <title>ol.start - reflection test</title>
  <link href="http://shiki.esrille.com/" rel="author" title="Shiki Okasaka"/>
  <link href="http://www.esrille.com/" rel="author" title="Esrille Inc."/>
  <link href="https://html.spec.whatwg.org/multipage/#the-ol-element" rel="help"/>
  <meta content="This test checks that the start IDL attribute reflects the respective content attribute of the same name." name="assert"/>
  <script src="/resources/testharness.js"/>
  <script src="/resources/testharnessreport.js"/>
 </head>
 <body>
  <ol id="ol">
   <li>One</li>
   <li>Two</li>
   <li>Three</li>
  </ol>
  <div id="log"/>
  <script>
test(function() {
    assert_equals(document.getElementById('ol').start, 1);
})
  </script>
 </body>
</html>
